Implemented functionality to delete authors via sessions
The ability to delete existing authors from the list has been enabled. When a user clicks the delete icon next to an author on the authors.php page, that author's ID is sent to authorDelete.php. This page accesses the session, finds the author with the corresponding ID, and removes them from the array of authors stored in $_SESSION. After successful deletion (or failure), the user is redirected back to authors.php with the updated list of authors